requirements.txt
fastapi
uvicorn[standard]
opencv-python-headless     # no GUI libs needed
pyyaml
inference                  # roboflow-inference-sdk
